feat: add utils and libs
