+++ src/components/pages/SinglePostPage.tsx
@@ -5,7 +5,7 @@ import gfm from "remark-gfm";
 import ReactMarkdown from "react-markdown";
 import { Toaster } from "react-hot-toast";
 
-import { db } from "../../firebase";
+import { db } from "../../libs/firebase";
 import { index as CodeBlock } from "../atom/code/index";
 import { index as Title } from "../atom/title/index";
 import { index as Loading } from "../atom/loading/index";
